refactor: sort out all the util methods into Util
