+++ src/chart/chart.js
@@ -7,7 +7,7 @@ const Util = require('../util');
 const View = require('./view');
 const G = require('../renderer2d');
 const Canvas = G.Canvas;
-const DomUtil = G.DomUtil;
+const DomUtil = Util.DomUtil;
 const Plot = require('../component/plot');
 const Controller = require('./controller/index');
 const Global = require('../global');
