+++ src/component/axis/circle.js
@@ -4,7 +4,7 @@
  */
 const Util = require('../../util');
 const Base = require('./base');
-const { vec2 } = require('../../renderer2d').MatrixUtil;
+const { vec2 } = Util.MatrixUtil;
 
 class Circle extends Base {
   getDefaultCfg() {
