+++ src/component/axis/line.js
@@ -4,7 +4,7 @@
  */
 const Base = require('./base');
 const Util = require('../../util');
-const { MatrixUtil } = require('../../renderer2d');
+const MatrixUtil = Util.MatrixUtil;
 const vec2 = MatrixUtil.vec2;
 
 class Line extends Base {
