+++ src/component/legend/category.js
@@ -4,7 +4,8 @@
  */
 const Util = require('../../util');
 const Base = require('./base');
-const { DomUtil, Event, Group } = require('../../renderer2d');
+const DomUtil = Util.DomUtil;
+const { Event, Group } = require('../../renderer2d');
 const Global = require('../../global');
 
 const CONTAINER_CLASS = 'g2-legend';
