+++ src/coord/polar.js
@@ -4,7 +4,7 @@
  */
 const Util = require('../util');
 const Base = require('./base');
-const MatrixUtil = require('../renderer2d').MatrixUtil;
+const MatrixUtil = Util.MatrixUtil;
 const mat3 = MatrixUtil.mat3;
 const vec2 = MatrixUtil.vec2;
 const vec3 = MatrixUtil.vec3;
