+++ src/geom/util/spline.js
@@ -1,4 +1,4 @@
-const MatrixUtil = require('../../renderer2d').MatrixUtil;
+const MatrixUtil = require('../../util').MatrixUtil;
 const Vector2 = MatrixUtil.vec2;
 
 function smoothBezier(points, smooth, isLoop, constraint) {
