+++ src/interaction/base.js
@@ -1,6 +1,6 @@
 // const Global = require('../global');
 const Util = require('../util');
-const { DomUtil } = require('../renderer2d');
+const DomUtil = Util.DomUtil;
 const View = require('../chart/view');
 const G2 = require('../core.js');
